Changes based off compiler warnings
The following code changes were made based off the warnings when
declaring the following configuration option:

WARNING_CFLAGS = -Wall -Wextra -Weverything -Wno-unused-parameter
-Wno-objc-missing-property-synthesis -Wno-format-nonliteral
-Wno-custom-atomic-properties

+++ Classes/Library/External Libraries/Sockets/GCDAsyncSocketCipherNames.m
@@ -364,7 +364,7 @@ static const char *kMacNames[] = {
 {
 	SSLCipherSuite cipher = [self sslNegotiatedCipherSuite];
 
-	for (NSInteger pos = 0; pos < sizeof(kCipherSuites) / sizeof(CipherSuite); pos++) {
+	for (unsigned long pos = 0; pos < sizeof(kCipherSuites) / sizeof(CipherSuite); pos++) {
 		CipherSuite cs = kCipherSuites[pos];
 
 		if (cs.cipher_suite == cipher) {
